+++ NEWS.md
@@ -55,6 +55,7 @@
 - Add a new `log_progress` argument to `make()` to suppress progress logging. This increases storage efficiency and speeds some projects up a tiny bit.
 - Add an optional `namespace` argument to `loadd()` and `readd()`. You can now load and read from non-default `storr` namespaces.
 - Add `drake_cache_log()`, `drake_cache_log_file()`, and `make(..., cache_log_file = TRUE)` as options to track changes to targets/imports in the drake cache.
+- Detect knitr code chunk dependencies in response to commands with `rmarkdown::render()`, not just `knit()`.
 
 # Version 4.4.0